+++ utils.py
@@ -1,0 +1,4 @@
+
+def raise_noob_readable_exception(exception_string):
+	# Prints a much more noticable exception for people to read.
+    raise Exception(f"\n\n==============================\n!!!HEY YOU, READ THIS!!!\n==============================\n{exception_string}")
